rename logprep.event_generator to logprep.generator
