+++ logprep/generator/kafka/document_sender.py
@@ -5,8 +5,8 @@ from logging import Logger
 from time import perf_counter
 from uuid import uuid4
 
-from logprep.event_generator.kafka.configuration import Configuration
-from logprep.event_generator.kafka.kafka_connector import KafkaProducer
+from logprep.generator.kafka.configuration import Configuration
+from logprep.generator.kafka.kafka_connector import KafkaProducer
 
 
 class DocumentSender:
