+++ tests/testdata/generator/kafka/config.yml
