+++ tests/testdata/generator/kafka/wineventlog_raw.jsonl
